Add files via upload
My first and last books in Russian language republished (1st) and published (volume 4) on Barnes and Noble in 2020-2021. The title "The Girl from California" (Девушка из штата Калифорния) has been written since 1995, when a teacher of English language from Russia (born in USSR) came to California, USA, and, by chance, got married to an American after the three weeks of her stay. The book is a never-ending story of a Russian educator who is in a constant journey through her work and studies, and who never stops wondering about the world and its people.  The observations are in her stories and jokes. Some stories have portraits of live people who will never know they have been portrayed,  Cultural and language barriers are not easy to overcome even in the computer age!